tsm-desktop/requirements.txt
textual
grpcio
grpcio-tools
yubikey-manager
